+++ custom_components/recalbox/intent.py
@@ -1,8 +1,10 @@
 from homeassistant.helpers import intent
+from homeassistant.core import HomeAssistant, State
 import unicodedata
 import re
-import homeassistant.helpers.config_validation as cv
 from .const import DOMAIN
+from .translations_service import RecalboxTranslator
+from .switch import RecalboxEntityMQTT
 import logging
 
 _LOGGER = logging.getLogger(__name__)
@@ -10,12 +12,16 @@ _LOGGER = logging.getLogger(__name__)
 async def async_setup_intents(hass):
     """Enregistre les handlers d'intentions seulement s'ils n'existent pas."""
 
+    if "intents_registered" in hass.data[DOMAIN]:
+        _LOGGER.debug("The intents are already registered. Skip re-register.")
+        return
+
     # Liste des intentions de votre intégration
     intents_to_register = [
         RecalboxLaunchHandler(),
         RecalboxStatusHandler(),
-        RecalboxActionHandler("RecalboxStopGame", 55355, "QUIT", "intent_response.quit_game_requested"),
-        RecalboxActionHandler("RecalboxPauseGame", 55355, "PAUSE_TOGGLE", "intent_response.pause_game_requested"),
+        RecalboxQuitGameHandler(),
+        RecalboxPauseGameHandler(),
         RecalboxScreenshotHandler()
     ]
 
@@ -23,35 +29,94 @@ async def async_setup_intents(hass):
         # On vérifie si l'intent_type est déjà enregistré
         if handler.intent_type not in intent.async_get(hass):
             intent.async_register(hass, handler)
+            _LOGGER.info(f"Registered {handler.intent_type} intent handler")
+
+    hass.data[DOMAIN]["intents_registered"] = True
 
 
-class RecalboxActionHandler(intent.IntentHandler):
-    def __init__(self, intent_type, port, command, responseKey):
-        self.intent_type = intent_type
-        self._port = port
-        self._command = command
-        self._responseKey = responseKey
+# --- TOOLS ----
 
-    async def async_handle(self, intent_obj):
-        instances = intent_obj.hass.data[DOMAIN].get("instances", {})
-        entry_id = list(instances.keys())[0]
-        api = instances[entry_id]["api"]
-        await api.send_udp_command(self._port, self._command)
-        translator = intent_obj.hass.data[DOMAIN]["translator"]
+# Va chercher la Recalbox par défaut.
+# - soit l'instance est spécifiée dans l'intent
+# - soit non, et on prend la première qu'on trouve allumée
+# - soit on prend la première qu'on trouve si toutes sont éteintes
+def find_recalbox_entity(hass: HomeAssistant, intent_obj:intent.Intent) -> RecalboxEntityMQTT:
+    instances = hass.data[DOMAIN].get("instances", {})
+    if not instances:
+        return None
 
-        response = intent_obj.create_response()
-        response.async_set_speech(translator.translate(self._responseKey, lang=intent_obj.language))
-        return response
+    # Extraction des entités réelles (RecalboxEntityMQTT)
+    all_entities = [instance.get("sensor_entity") for instance in instances.values() if instance.get("sensor_entity")]
+
+    if not all_entities:
+        _LOGGER.warning("No RecalboxEntityMQTT found")
+        return None
+
+    # Prio 1 :
+    # Recherche par nom spécifique, si présent dans l'Intent
+    slots = intent_obj.slots
+    recalbox_instance_name = slots.get("instance", {}).get("value")
+
+    if recalbox_instance_name:
+        target_slug = slugify(str(recalbox_instance_name))
+        for entity in all_entities:
+            # On compare avec le nom de l'entité ou un attribut de nom
+            if slugify(entity.name) == target_slug:
+                _LOGGER.info(f"Found Recalbox target by its name {entity.name}")
+                return entity
+
+    # Prio 2 :
+    # Si on n'a pas trouvé un RecalboxEntityMQTT par nom,
+    # on cherche la première qui est "ON"
+    for entity in all_entities:
+        state = hass.states.get(entity.entity_id)
+        if state and state.state == "on":
+            _LOGGER.info(f"Found a Recalbox that is ON : {entity.name}")
+            return entity
+
+    # Sinon, on prend juste la première de la liste par défaut
+    _LOGGER.info(f"Using the first Recalbox by default.")
+    return all_entities[0]
+
+def find_recalbox_states(hass: HomeAssistant, intent_obj:intent.Intent) -> State:
+    recalboxEntity:RecalboxEntityMQTT = find_recalbox_entity(hass, intent_obj)
+    return hass.states.get(recalboxEntity.entity_id)
+
+def get_translator(hass: HomeAssistant) -> RecalboxTranslator:
+    return hass.data[DOMAIN]["translator"]
+
+
+
+# ---- Intent Handlers -----
+
+
+
+#class RecalboxOtherUDPActionHandler(intent.IntentHandler):
+#    def __init__(self, intent_type, port, command, responseKey):
+#        self.intent_type = intent_type
+#        self._port = port
+#        self._command = command
+#        self._responseKey = responseKey
+#
+#    async def async_handle(self, intent_obj):
+#        instances = intent_obj.hass.data[DOMAIN].get("instances", {})
+#        entry_id = list(instances.keys())[0]
+#        api = instances[entry_id]["api"]
+#        await api.send_udp_command(self._port, self._command)
+#        translator = intent_obj.hass.data[DOMAIN]["translator"]
+#
+#        response = intent_obj.create_response()
+#        response.async_set_speech(translator.translate(self._responseKey, lang=intent_obj.language))
+#        return response
+
 
 class RecalboxScreenshotHandler(intent.IntentHandler):
     intent_type = "RecalboxCreateScreenshot"
 
     async def async_handle(self, intent_obj):
         hass = intent_obj.hass
-        instances = hass.data[DOMAIN].get("instances", {})
-        entry_id = list(instances.keys())[0]
-        recalbox = instances[entry_id].get("sensor_entity")
-        translator = hass.data[DOMAIN]["translator"]
+        recalbox:RecalboxEntityMQTT = find_recalbox_entity(hass, intent_obj)
+        translator:RecalboxTranslator = get_translator(hass)
 
         if await recalbox.request_screenshot():
             text = translator.translate("intent_response.screenshot_success", lang=intent_obj.language)
@@ -62,26 +127,63 @@ class RecalboxScreenshotHandler(intent.IntentHandler):
         response.async_set_speech(text)
         return response
 
+
+
+class RecalboxQuitGameHandler(intent.IntentHandler):
+    intent_type = "RecalboxStopGame"
+
+    async def async_handle(self, intent_obj):
+        hass = intent_obj.hass
+        recalbox:RecalboxEntityMQTT = find_recalbox_entity(hass, intent_obj)
+        translator:RecalboxTranslator = get_translator(hass)
+
+        if await recalbox.request_quit_current_game():
+            text = translator.translate("intent_response.quit_game_requested", lang=intent_obj.language)
+        else:
+            text = translator.translate("intent_response.quit_game_failed", lang=intent_obj.language)
+
+        response = intent_obj.create_response()
+        response.async_set_speech(text)
+        return response
+
+
+
+class RecalboxPauseGameHandler(intent.IntentHandler):
+    intent_type = "RecalboxPauseGame"
+
+    async def async_handle(self, intent_obj):
+        hass = intent_obj.hass
+        recalbox:RecalboxEntityMQTT = find_recalbox_entity(hass, intent_obj)
+        translator:RecalboxTranslator = get_translator(hass)
+
+        if await recalbox.request_pause_game():
+            text = translator.translate("intent_response.pause_game_requested", lang=intent_obj.language)
+        else:
+            text = translator.translate("intent_response.pause_game_failed", lang=intent_obj.language)
+
+        response = intent_obj.create_response()
+        response.async_set_speech(text)
+        return response
+
+
+
 class RecalboxStatusHandler(intent.IntentHandler):
     intent_type = "RecalboxGameStatus"
 
     async def async_handle(self, intent_obj):
         # On va lire l'état de l'entité binary_sensor pour répondre
         hass = intent_obj.hass
-        instances = hass.data[DOMAIN].get("instances", {})
-        entry_id = list(instances.keys())[0]
-        recalboxEntity = instances[entry_id].get("sensor_entity")
-        recalbox = hass.states.get(recalboxEntity.entity_id)
-        translator = hass.data[DOMAIN]["translator"]
+        recalboxState:State = find_recalbox_states(hass, intent_obj)
+        translator:RecalboxTranslator = get_translator(hass)
 
-        if not recalbox:
+        if not recalboxState:
             text = translator.translate("intent_response.recalbox_not_found", lang=intent_obj.language)
-        elif recalbox.state == "off":
+        elif recalboxState.state == "off":
             text = translator.translate("intent_response.recalbox_offline", lang=intent_obj.language)
         else:
-            game = recalbox.attributes.get("game", "-")
+            game = recalboxState.attributes.get("game", "-")
             if game is not None and game != "None" and game != "-" :
-                console = recalbox.attributes.get("console", "")
+                console = recalboxState.attributes.get("console", "")
                 text = translator.translate(
                     "intent_response.game_status_playing",
                     {"game": game, "console": console},
@@ -99,16 +201,15 @@ class RecalboxLaunchHandler(intent.IntentHandler):
     """Handler pour lancer un jeu."""
     intent_type = "RecalboxLaunchGame" # Doit correspondre au nom dans ton YAML
 
-    async def async_handle(self, intent_obj):
+    async def async_handle(self, intent_obj:intent.Intent):
         hass = intent_obj.hass
-        instances = hass.data[DOMAIN].get("instances", {})
-        entry_id = list(instances.keys())[0]
-        recalbox = instances[entry_id].get("sensor_entity")
-
         # 1. Récupérer les slots (variables) de la phrase
         slots = intent_obj.slots
         game = slots.get("game", {}).get("value")
         console = slots.get("console", {}).get("value")
+        # recalboxEntityId = slots.get("recalboxEntityId", {}).get("value")
+
+        recalbox = find_recalbox_entity(hass, intent_obj)
 
         # Appeler la fonction de recherche
         result_text = await recalbox.search_and_launch_game_by_name(console, game, lang=intent_obj.language)