+++ custom_components/recalbox/switch.py
@@ -2,7 +2,9 @@ from homeassistant.components.mqtt import async_subscribe
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed, CoordinatorEntity
 from homeassistant.helpers import device_registry as dr
+from homeassistant.exceptions import HomeAssistantError
 from .const import DOMAIN
+from .translations_service import RecalboxTranslator
 import unicodedata
 import re
 import homeassistant.helpers.config_validation as cv
@@ -126,7 +128,9 @@ class RecalboxEntityMQTT(CoordinatorEntity, SwitchEntity):
 
 
     async def async_turn_on(self, **kwargs):
-        _LOGGER.error("L'allumage de la Recalbox n'est pas supporté à distance.")
+        translator:RecalboxTranslator = self.hass.data[DOMAIN]["translator"]
+        power_off_not_implemented_message = translator.translate("errors.power_off_not_implemented_message")
+        raise HomeAssistantError(power_off_not_implemented_message)
 
 
     #################################
@@ -137,12 +141,6 @@ class RecalboxEntityMQTT(CoordinatorEntity, SwitchEntity):
         _LOGGER.debug("Forcing Recalbox status OFF (sans attendre MQTT)")
         self._attr_is_on = False
         self.async_write_ha_state()
-
-
-    # Exemple : Action appelée par un service ou un bouton
-    async def request_turn_on(self, **kwargs) -> bool:
-        """Appelé quand on clique sur ON dans l'interface."""
-        return await self.send_udp_command("POWER_ON")
 
 
     async def request_shutdown(self) -> bool:
@@ -200,7 +198,7 @@ class RecalboxEntityMQTT(CoordinatorEntity, SwitchEntity):
     # Renvoie le texte pour Assist
     async def search_and_launch_game_by_name(self, console, game_query, lang=None) -> str :
         _LOGGER.debug(f"Try to launch game {game_query} on system {console}")
-        translator = self.hass.data[DOMAIN]["translator"]
+        translator:RecalboxTranslator = self.hass.data[DOMAIN]["translator"]
         # Récupérer la liste des roms via l'API (HTTP GET)
         try:
             roms = await self._api.get_roms(console)