+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 > By Aurélien Tomassini, 2026.
 
 
+## v1.2.1 - 26/01/2026
+
+
+- In the dashboard card, remove the interactive switch, to a simple power state icon
+- Add an exception when trying to turn on Recalbox from HA (only turn off is supported)
+- Externalize the custom_sentences installer to be an sync job
+- Externalize Recalbox services installation
+- Improve code and logs on assist sentences and intents registration
+- For Voice/Text assis, now looks for the Recalbox that are ON, instead of using the 1st one by default.
+  Preparing also commands to specify a Recalbox target by name.
+- Improve french pausing/resuming sentences
+
+
+
 ## v1.2.0 - 26/01/2026 Breaking change
 
 > This versions changes the entity from `binary sensor` to `switch`.